01/06 :recycle: : user Profile 컴포넌트 분리
